Add files via upload
Resorted to millisec for this and it seems to work well, I can use it in gated mode or retrigger mode by monitoring the mono/multi setting from the synth and watching for key triggers on the digital pins. Delay from 0 - 10 seconds controlled by the synth CV with retriggering and +5v trigger to drive the env generator for slope.